Move SetupPassword submit logic to separate hook
